requirements.txt
arrow==0.17.0
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
coverage==5.5
Django==3.1.7
django-tinymce==3.2.0
humanize==3.2.0
idna==2.10
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-time==0.3.0
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.3
views==0.3
fontawesome-free==5.15.3
sorl-thumbnail==12.7.0
Pillow==8.1.2
gunicorn==20.1.0
requests==2.25.1
requests-toolbelt==0.9.1
dj-database-url==0.5.0
